Backend/requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Deprecated==1.2.13
fastapi==0.95.1
filelock==3.11.0
firebase-admin==6.1.0
fonttools==4.39.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
jwcrypto==1.4.2
kiwisolver==1.4.4
lit==16.0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
thop==0.1.1.post2209072238
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.88
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
yapf==0.33.0
